refactor : 내 그룹 조회 페이징 수정
- 내 그룹 조회 페이징 수정
- 테스트 코드 추가

+++ src/main/java/com/sosim/server/group/GroupRepositoryDsl.java
@@ -3,5 +3,5 @@ package com.sosim.server.group;
 import org.springframework.data.domain.Slice;
 
 public interface GroupRepositoryDsl {
-    public Slice<Group> findMyGroups(long userId, int offset, int limit);
+    public Slice<Group> findMyGroups(long userId, long offset, long limit);
 }
