+++ src/main/java/com/sosim/server/group/MyGroupPaginationUtil.java
@@ -3,12 +3,12 @@ package com.sosim.server.group;
 import com.sosim.server.group.dto.MyGroupPageDto;
 
 public class MyGroupPaginationUtil {
-    private static final int FIRST_SIZE = 17;
-    private static final int OTHER_SIZE = 18;
+    private static final long FIRST_SIZE = 17;
+    private static final long OTHER_SIZE = 18;
 
     public static MyGroupPageDto calculateOffsetAndSize(int page) {
-        int offset = calculateOffset(page);
-        int limit = isFirstPage(page) ? FIRST_SIZE : OTHER_SIZE;
+        long offset = calculateOffset(page);
+        long limit = isFirstPage(page) ? FIRST_SIZE : OTHER_SIZE;
 
         return new MyGroupPageDto(offset, limit);
     }
@@ -17,8 +17,8 @@ public class MyGroupPaginationUtil {
         return page == 0;
     }
 
-    private static int calculateOffset(int page) {
-        int offset = (page - 1) * OTHER_SIZE + FIRST_SIZE;
+    private static long calculateOffset(int page) {
+        long offset = (page - 1) * OTHER_SIZE + FIRST_SIZE;
         if (offset < 0) {
             offset = 0;
         }
